pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "telegram>=0.0.1",
]
